OPTIMAL BOGGLE ADDED
Optimal boggle is a revision of the previous boggle by using recursion, it proves to be quite fast but still not as efficient.

Problem still lies within the list itself, next revision will be weeding out the unwanted from the master list.

Also a competition between the Boyer-Moore-Horspool algorithm and Boggle will occur.